refactor: extract pure session and settings helpers for tests
Move snapshot parse and stale deleteGuildIds filtering into pure modules,
and expose clear-skip state helpers so regressions can be covered without DB.
